Check HTTP methods

+++ backend/src/com/carlgo11/tempfiles/api/Cleanup.php
@@ -3,11 +3,14 @@
 namespace com\carlgo11\tempfiles\api;
 
 use com\carlgo11\tempfiles\DataStorage;
+use Exception;
 
 class Cleanup extends API
 {
 
-    public function __construct() {
+    public function __construct(string $method) {
+        if ($method !== 'PURGE') throw new Exception("Bad HTTP method. Use PURGE.");
+
         $status = filter_var(DataStorage::deleteOldFiles(), FILTER_VALIDATE_BOOLEAN, FILTER_NULL_ON_FAILURE);
         parent::addMessage('status', $status);
         parent::outputJSON(202);
